requirements.txt
bottle 0.12.7
beaker 1.6.3
sqlalchemy 0.9.0
requests 2.3.0
python-lxml 3.4.0
python-cssselect 0.9.1
